change user_id from int to long

+++ restful-web-service/src/main/java/com/pmp/restful_web_service/controller/PostController.java
@@ -35,7 +35,7 @@ public class PostController {
      * @return List<Post>
      */
     @GetMapping("user/{userId}")
-    public List<Post> getAllPostsForUser(@PathVariable int userId) {
+    public List<Post> getAllPostsForUser(@PathVariable long userId) {
         return this.postService.getAllPostsForUser(userId);
     }
